#1559 - Problems if multiple entities have the same label
- Pull code which is only required for the concept feature editor into the that one creating an inner class with additional modifications of the KnowledgeBaseItemAutoCompleteField
- Leave generic code not in the KnowledgeBaseItemAutoCompleteField so that it can be used on the KB page for the search field and for the IRI value editor
- Removed unused dependency
