+++ inception-ui-kb/src/main/java/de/tudarmstadt/ukp/inception/ui/kb/feature/KnowledgeBaseItemAutoCompleteField.java
@@ -21,15 +21,9 @@ import static org.apache.wicket.RuntimeConfigurationType.DEVELOPMENT;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
 
 import org.apache.commons.lang3.Validate;
-import org.apache.wicket.ajax.form.AjaxFormChoiceComponentUpdatingBehavior;
 import org.apache.wicket.model.IModel;
-import org.apache.wicket.request.IRequestParameters;
-import org.apache.wicket.request.http.WebRequest;
-import org.apache.wicket.util.convert.IConverter;
-import org.apache.wicket.util.string.StringValue;
 import org.danekja.java.util.function.serializable.SerializableFunction;
 
 import com.googlecode.wicket.jquery.core.JQueryBehavior;
@@ -43,11 +37,6 @@ import de.tudarmstadt.ukp.inception.kb.graph.KBHandle;
 
 /**
  * Auto-complete field for accessing a knowledge base.
- * <p>
- * <b>NOTE:</b> This field <b>must</b> be used with a
- * {@link AjaxFormChoiceComponentUpdatingBehavior} and using
- * {@link #getIdentifierDynamicAttributeScript()} otherwise it cannot differentiate between
- * different items that have the same label!!!
  */
 public class KnowledgeBaseItemAutoCompleteField
     extends AutoCompleteTextField<KBHandle>
@@ -55,9 +44,6 @@ public class KnowledgeBaseItemAutoCompleteField
     private static final long serialVersionUID = -1276017349261462683L;
 
     private SerializableFunction<String, List<KBHandle>> choiceProvider;
-    private IConverter<KBHandle> converter;
-    private List<KBHandle> choiceCache;
-    private boolean allowChoiceCache = false;
     
     public KnowledgeBaseItemAutoCompleteField(String aId,
             SerializableFunction<String, List<KBHandle>> aChoiceProvider)
@@ -66,7 +52,6 @@ public class KnowledgeBaseItemAutoCompleteField
         
         Validate.notNull(aChoiceProvider);
         choiceProvider = aChoiceProvider;
-        converter = newConverter();
     }
 
     public KnowledgeBaseItemAutoCompleteField(String aId,
@@ -77,7 +62,6 @@ public class KnowledgeBaseItemAutoCompleteField
         
         Validate.notNull(aChoiceProvider);
         choiceProvider = aChoiceProvider;
-        converter = newConverter();
     }
 
     public KnowledgeBaseItemAutoCompleteField(String aId,
@@ -88,124 +72,14 @@ public class KnowledgeBaseItemAutoCompleteField
         
         Validate.notNull(aChoiceProvider);
         choiceProvider = aChoiceProvider;
-        converter = newConverter();
     }
     
     @Override
     protected List<KBHandle> getChoices(String aInput)
     {
-        if (!allowChoiceCache || choiceCache == null) {
-            choiceCache = choiceProvider.apply(aInput);
-        }
-        return choiceCache;
+        return choiceProvider.apply(aInput);
     }
     
-    @Override
-    public String[] getInputAsArray()
-    {
-        // If the web request includes the additional "identifier" parameter which is supposed to
-        // contain the IRI of the selected item instead of its label, then we use that as the value.
-        WebRequest request = getWebRequest();
-        IRequestParameters requestParameters = request.getRequestParameters();
-        StringValue identifier = requestParameters
-                .getParameterValue(getInputName() + ":identifier");
-        
-        if (!identifier.isEmpty()) {
-            return new String[] { identifier.toString() };
-        }
-        
-        return super.getInputAsArray();
-    }
-    
-    /**
-     * When using this input component with an {@link AjaxFormChoiceComponentUpdatingBehavior}, it
-     * is necessary to request the identifier of the selected item as an additional dynamic 
-     * attribute, otherwise no distinction can be made between two items with the same label!
-     */
-    public String getIdentifierDynamicAttributeScript()
-    {
-        return String.join(" ", 
-                "var item = $(attrs.event.target).data('kendoAutoComplete').dataItem();",
-                "if (item) {",
-                "  return [{",
-                "    'name': '" + getInputName() + ":identifier', ",
-                "    'value': $(attrs.event.target).data('kendoAutoComplete').dataItem().identifier",
-                "  }]",
-                "}",
-                "return [];");
-    }
-    
-    @Override
-    public <C> IConverter<C> getConverter(Class<C> aType)
-    {
-        if (aType != null && aType.isAssignableFrom(this.getType())) {
-            return (IConverter<C>) converter;
-        }
-        
-        return super.getConverter(aType);
-    }
-    
-    private IConverter<KBHandle> newConverter()
-    {
-        return new IConverter<KBHandle>() {
-
-            private static final long serialVersionUID = 1L;
-
-            @Override
-            public KBHandle convertToObject(String value, Locale locale)
-            {
-                if (value == null) {
-                    return null;
-                }
-                
-                if (value.equals(getModelValue())) {
-                    return getModelObject();
-                }
-                
-                // Check choices only here since fetching choices can take some time. If we already
-                // have choices from a previous query, then we use them instead of reloading all
-                // the choices. This avoids having to load the choices when opening the dropdown
-                // AND when selecting one of the items from it.
-                List<KBHandle> choices; 
-                try {
-                    allowChoiceCache = true;
-                    choices = getChoices(value);
-                }
-                finally {
-                    allowChoiceCache = false;
-                }
-                
-                if (choices.isEmpty()) {
-                    return null;
-                }
-                
-                // Check if we can find a match by the identifier. The identifier is unique while
-                // the same label may appear on multiple items
-                for (KBHandle handle : choices) {
-                    if (value.equals(handle.getIdentifier())) {
-                        return handle;
-                    }
-                }
-                
-                // Check labels if there was no match on the identifier
-                for (KBHandle handle : choices) {
-                    if (value.equals(getRenderer().getText(handle))) {
-                        return handle;
-                    }
-                }
-                
-                // If there was no match at all, return null
-                return null;
-            }
-
-            @Override
-            public String convertToString(KBHandle value, Locale locale)
-            {
-                return getRenderer().getText(value);
-            }
-        };
-    } 
-
     @Override
     public void onConfigure(JQueryBehavior behavior)
     {
@@ -229,15 +103,6 @@ public class KnowledgeBaseItemAutoCompleteField
         behavior.setOption("filtering", String.join(" ",
                 "function(e) {",
                 "  e.sender.listView.value([]);",
-                "}"));
-        
-        // We need to explicitly trigger the change event on the input element in order to trigger
-        // the Wicket AJAX update (if there is one). If we do not do this, then Kendo will "forget"
-        // to trigger a change event if the label of the newly selected item is the same as the 
-        // label of the previously selected item!!!
-        behavior.setOption("select", String.join(" ",
-                "function (e) {",
-                "  e.sender.element.trigger('change');",
                 "}"));
         
         // Prevent scrolling action from closing the dropdown while the focus is on the input field
